change level type | rearrange Staked event args

+++ contracts/Types.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.27;
 
 struct StakingData {
-    uint256 level;
+    uint8 level;
     uint256 amount;
     uint256 lockEndTime;
     uint256 vehicleId;
